Extract IntegrationTestCase base class for future use
