+++ tests/test_simpleflow/process/test_supervisor.py
@@ -1,44 +1,17 @@
 import multiprocessing
 import os
-import re
 import signal
 import time
-import unittest
 
-from psutil import Process, NoSuchProcess
+from psutil import Process
 from setproctitle import setproctitle
 from sure import expect
 
 from simpleflow.process import Supervisor, reset_signal_handlers
+from tests.utils import IntegrationTestCase
 
 
-class TestSupervisor(unittest.TestCase):
-    def setUp(self):
-        pass
-
-    def tearDown(self):
-        # cleanup all child processes
-        if os.getenv("SIMPLEFLOW_CLEANUP_PROCESSES"):
-            process = Process()
-            for child in process.children(recursive=True):
-                # TODO: have a warning here? normal?
-                try:
-                    child.kill()
-                except NoSuchProcess:
-                    pass
-        # reset SIGTERM handler
-        signal.signal(signal.SIGTERM, signal.SIG_DFL)
-
-    def assertProcess(self, regex, count=1):
-        children = Process().children(recursive=True)
-        matching = [p for p in children if re.search(regex, p.name())]
-        self.assertEqual(
-            len(matching), count,
-            "Expected {} processes matching {}, found {} in {}.".format(
-                count, regex, len(matching), children
-            )
-        )
-
+class TestSupervisor(IntegrationTestCase):
     def test_start(self):
         # dummy function used in following tests
         def sleep_long(seconds):
@@ -117,7 +90,7 @@ class TestSupervisor(unittest.TestCase):
         expect(len(old_workers)).to.equal(1)
 
         # now kill the worker
-        worker_pid = old_workers[0].pid
+        old_workers[0].pid
         os.kill(workers()[0].pid, signal.SIGKILL)
         time.sleep(0.5)
 
@@ -129,6 +102,7 @@ class TestSupervisor(unittest.TestCase):
     # NB: not in the Supervisor class but we want to benefit from the tearDown()
     def test_reset_signal_handlers(self):
         signal.signal(signal.SIGTERM, signal.SIG_IGN)
+
         def foo():
             time.sleep(1)
